+++ .github/workflows/linters.yml
@@ -17,4 +17,4 @@ jobs:
           gem install --no-document rubocop -v '>= 1.0, < 2.0' # https://docs.rubocop.org/en/stable/installation/
           [ -f .rubocop.yml ] || wget https://raw.githubusercontent.com/microverseinc/linters-config/master/ruby/.rubocop.yml
       - name: Rubocop Report
-        run: rubocop --color --fix
+        run: rubocop --color
