+++ README.md
@@ -2,13 +2,13 @@
 
 ![](https://img.shields.io/badge/Microverse-blueviolet)
 
-In this project, we have used SOLID principles of OOP, YAGNI, KISS, and DRY rules and Ruby style guide to refactor our School Library App.
+> This project is for creating a tool to record what books are in the library and who borrows them. It also add a feature of adding a book, student, and teacher.
 
 ## Built With
 - Ruby
 
 ## Getting Started
-1. ```git clone https://github.com/samiullahbahadur/school-library```
+1. ```git clone https://github.com/bushmusi/school-library```
 2. ```cd school-library```
 ### Prerequests
 1. [Git](https://git-scm.com/downloads).
@@ -32,30 +32,9 @@ rspec spec/operation_spec.rb
 rspec spec/group_spec.rb 
 ```
 
-## Author1
+## Author
 
-- GitHub: [@Bahadur](https://github.com/samiullahbahadur)
-- Twitter: [@Bahadur](https://twitter.com/Samiull88496331)
-- LinkedIn: [@Bahadur](https://www.linkedin.com/in/samiullah-bahadur/)
+- GitHub: [@bushra](https://github.com/bushmusi)
+- Twitter: [@bushra](https://twitter.com/bushera_mestofa)
+- LinkedIn: [@bushra](https://www.linkedin.com/in/bushra-mustofa)
 
-## Author2
-**Nicholas Mutale**
-
-- GitHub: [@mutalenic](https://github.com/Mutalenic)
-- Twitter: [@nicomutale](https://twitter.com/nicomutale)
-- LinkedIn: [@nicomutale](https://www.linkedin.com/in/nicomutale/)
-
-## Contributing
-
-Contributions, issues, and feature requests are welcome!
-
-Feel free to check the [issues page](https://github.com/Mutalenic/school_library/issues).
-
-## Show your support
-
-Give a ⭐️ if you like this project!
-
-
-## Acknowledgments
-
-- Microverse
