+++ app.rb
@@ -1,10 +1,8 @@
-require './books'
-
 class App
   def initialize
     @store = {
       persons: [],
-      @books = Books.new
+      books: [],
       rentals: []
     }
     @methods = {
@@ -54,7 +52,21 @@ class App
     puts 'Saved.'
   end
 
-  
+  def create_book
+    puts 'Please add details of the book'
+
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+
+    require './book'
+    the_book = Book.new(title, author)
+
+    @store[:books] << the_book
+    puts 'Saved.'
+  end
+
   def create_a_rental
     if @store[:books].length.zero? || @store[:persons].length.zero?
       puts 'Please make sure you have at least one person and one book in the database'
