+++ person.rb
@@ -6,7 +6,7 @@ class Person < Nameable
 
   attr_reader :id
 
-  def initialize(age, name = 'unknown', parent_permission = true)
+  def initialize(age, name = 'unknown', parent_permission: true)
     super()
     @id = Random.rand(0..1000)
     @name = name
